feat(main): split numericalAnalysis.ts into separate files
